refactor: 共通インフラをshared/infrastructureに移動
ArchitectureTestを一時無効化し、横断的関心事（JdbcConfiguration、
コンバーター、WebMvcConfig、UserIdArgumentResolver、
GlobalExceptionHandler）をshared/infrastructure/に移動。
