+++ backend/src/main/java/com/youmorry/expensetracker/shared/infrastructure/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.youmorry.expensetracker.shared.infrastructure;
+
+import org.jspecify.annotations.NullMarked;
