+++ backend/src/main/java/com/youmorry/expensetracker/shared/infrastructure/persistence/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.youmorry.expensetracker.shared.infrastructure.persistence;
+
+import org.jspecify.annotations.NullMarked;
