+++ backend/src/main/java/com/youmorry/expensetracker/shared/infrastructure/web/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.youmorry.expensetracker.shared.infrastructure.web;
+
+import org.jspecify.annotations.NullMarked;
