+++ backend/src/test/java/com/youmorry/expensetracker/ArchitectureTest.java.disabled
